RestServiceClientBase: List instead of Array
Previously, we deserialized arrays from (top level) json collections.
I mean, if egy we received smg like this: {someKey:[{...},...]}
It's not optimal, because e.g. so we cannot modify the user's carts(, without converting it to a List later). So, from now, we will deserialize Lists, not Arrays

Also done:
* RestServiceClientBase.java: Changed the DELETE HTTP method to work like a GET, so the data won't be placed into the request body, instead the query string will be used for this. Yeah, this was a very big mistake about the HTTP protocol, I wonder how WCF could let it so... And why it works that way (the wrong way, placing the data into the request body); and why not working yet the right way, as query string...