Extracted DevMode fns (AndroidClientProxy lib)
There were developer mode only parts in the code earlier as well. Now, they got extracted into a separate file/package. All of them will be disabled when the build is not in debug mode. Them can be disabled one-by-one via Config

The existing/extracted parts were:
* Applying mock data if there is no internet connection available

Note: There is a lot of code duplication around the developer mode's framework (they are the same, exactly). The benefit of this is, that the 2 module could have 2 separated DevMode section, independently from the other. But, of course, a 3. module (e.g. AndroidCommon) would be a better place to merge up this 2 frameworks, and only halt the module specific part in the existing modules.