HTTP DELETE: Temporary fix for WCF's misunderstanding of HTTP DELETE
For some reason, at least in this project, WCF misunderstands HTTP DELETE, handles it like POST (data in body), not like GET (data in query string). This also means, if you send it in a GET-like format, it won't work!
So, we have to send DELETE requests in POST-like format. It worked while implementing the WP8 client; but on Android(, using the Apache's HttpClient lib), you can't put data into the body by a DELETE request.
So, temporary just enabled also POST* by the DELETE required services; and on Android, we'll always send POST instead of DELETE requests
*more honestly, every HTTP verb

Also done:
* Enchanced the WCF log's maxSizeOfMessageToLog property